update: major frontend code cleanup
* Embed componentt refactoring
 * split parts to new components
 * handler cleanup
* CSS updates
