+++ ui/src/index.css
@@ -63,12 +63,12 @@ button:hover {
 }
 
 /* Hover styles for the delete button */
-button#delete-btn:hover {
+button.delete-btn:hover {
     background-color: red;
     color: white;
 }
 
-button#update-btn:hover {
+button.update-btn:hover {
     background-color: green;
     color: white;
 }
@@ -76,6 +76,11 @@ button#update-btn:hover {
 button:active {
     box-shadow: 0 0px 1px hsla(0, 0%, 0%, 0.4);
     transform: translateY(1px);
+}
+
+.echart-container {
+    height: 350px;
+    width: 350px;
 }
 
 .sr-only {
@@ -279,13 +284,13 @@ i {
     padding-bottom: 20px;
 }
 
-#embed-options {
+#embed-input {
     display: flex;
     flex-direction: column;
     gap: 1rem;
 }
 
-#embed-form-text-options {
+#embed-projection {
     display: flex;
     flex-direction: column;
     gap: 1rem;
